requirements.txt
# Stripe Support Agent - Python Dependencies

# Configuration management
pydantic-settings>=2.0.0

# Text processing
tiktoken>=0.5.0
beautifulsoup4>=4.12.0

# Embeddings and vector storage
sentence-transformers>=2.2.0
qdrant-client>=1.7.0

# Utilities
tqdm>=4.65.0

# Web scraping and HTTP
crawl4ai>=0.2.0
aiohttp>=3.9.0
httpx>=0.25.0

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# AI/LLM
anthropic>=0.18.0
langgraph>=0.0.40

# Search
rank_bm25>=0.2.2

# Observability
langfuse>=2.0.0
